feat: now prints json report with argument if supplied. cleaned up code
