+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 dependency-status-report.html
+dependency-status-report.json
 yarn-error.log
