+++ src/enums.js
@@ -1,0 +1,8 @@
+export const reportTypes = {
+  JSON: "JSON",
+  HTML: "HTML",
+};
+
+export const STATUS_UP_TO_DATE = "UP TO DATE";
+export const STATUS_OUTDATED = "OUTDATED";
+export const STATUS_UNKNOWN = "UNKNOWN";
